+++ src/components/settings/profile.js
@@ -1,36 +1,15 @@
 import React from 'react';
 import {
-  StyleSheet, View, Text, TextInput, Button,
+  StyleSheet, View,
 } from 'react-native';
+import { TextField } from '../../shared/UI/inputs';
+import { SubmitButton } from '../../shared/UI/buttons';
 import i18n from '../../shared/i18n';
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-  },
-  formItem: {
-    marginHorizontal: 16,
-    marginVertical: 8,
-  },
-  label: {
-    fontSize: 18,
-    color: 'black',
-    marginBottom: 2,
-  },
-  input: {
-    fontSize: 16,
-    borderWidth: 1,
-    borderRadius: 3,
-    backgroundColor: 'white',
-    textAlignVertical: 'top',
-    padding: 10,
-  },
-  inputName: { },
-  inputDescription: {
-    height: 100,
-  },
-  button: {
-    marginTop: 5,
+    justifyContent: 'flex-start',
   },
 });
 
@@ -38,38 +17,20 @@ const formScope = { scope: 'formFields' };
 
 const Profile = ({
   deviceName, onChangeDeviceName, onSubmit,
-}) => {
-  const deviceNameInput = (
-    <View style={styles.formItem}>
-      <Text style={styles.label}>
-        {i18n.t('deviceName', formScope)}
-      </Text>
-      <TextInput
-        style={[styles.input, styles.inputName]}
-        value={deviceName}
-        placeholder={i18n.t('shortName', formScope)}
-        onChangeText={onChangeDeviceName}
-      />
-    </View>
-  );
-
-  const submitText = i18n.t('save');
-  const submitButton = (
-    <View style={[styles.formItem, styles.button]}>
-      <Button
-        title={submitText}
-        onPress={onSubmit}
-      />
-    </View>
-  );
-
-  return (
-    <View style={styles.container}>
-      {deviceNameInput}
-      {submitButton}
-    </View>
-  );
-};
+}) => (
+  <View style={styles.container}>
+    <TextField
+      label={i18n.t('deviceName', formScope)}
+      value={deviceName}
+      placeholder={i18n.t('shortName', formScope)}
+      onChange={onChangeDeviceName}
+    />
+    <SubmitButton
+      text={i18n.t('save')}
+      onPress={onSubmit}
+    />
+  </View>
+);
 
 
 export default Profile;