+++ src/containers/subjects/form.js
@@ -34,17 +34,23 @@ export class SubjectForm extends React.Component {
   constructor(props) {
     super(props);
 
-    const { name, description, category } = this.props.subject || {};
+    const {
+      name, description, category, icon,
+    } = this.props.subject || {};
 
     this.state = {
       name,
       description,
       categoryId: category ? category.id : -1,
+      icon,
     };
 
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmitSubject = this.handleSubmitSubject.bind(this);
+    this.handleChangeName = this.handleChange('name');
+    this.handleChangeDescription = this.handleChange('description');
+    this.handleChangeCategory = this.handleChange('categoryId');
+    this.handleChangeIcon = this.handleChange('icon');
 
+    this.handleSubmitSubject = this.handleSubmitSubject.bind(this);
     this.handleDeleteSubject = this.handleDeleteSubject.bind(this);
 
     this.props.navigation.setParams({
@@ -68,12 +74,15 @@ export class SubjectForm extends React.Component {
       return;
     }
 
-    const { name, description, categoryId } = this.state;
+    const {
+      name, description, categoryId, icon,
+    } = this.state;
 
     const data = {
       name: name.trim(),
       description: description && description.trim(),
       category: categoryId,
+      icon,
     };
 
     if (this.props.subject) {
@@ -109,10 +118,20 @@ export class SubjectForm extends React.Component {
     });
   }
 
+  findCategoryColor() {
+    const { categories } = this.props;
+    const { categoryId } = this.state;
+    const category = (categories || []).find(cat => cat.id === categoryId);
+    return category && category.color;
+  }
+
   render() {
-    const { name, description, categoryId } = this.state;
+    const {
+      name, description, categoryId, icon,
+    } = this.state;
     const { categories } = this.props;
     const canSubmit = this.isInputValid();
+    const color = this.findCategoryColor();
 
     return (
       <SubjectFormComponent
@@ -120,9 +139,12 @@ export class SubjectForm extends React.Component {
         description={description}
         categoryId={categoryId}
         categories={categories}
-        onChangeName={this.handleChange('name')}
-        onChangeDescription={this.handleChange('description')}
-        onChangeCategory={this.handleChange('categoryId')}
+        icon={icon}
+        color={color}
+        onChangeName={this.handleChangeName}
+        onChangeDescription={this.handleChangeDescription}
+        onChangeCategory={this.handleChangeCategory}
+        onChangeIcon={this.handleChangeIcon}
         onSubmit={this.handleSubmitSubject}
         canSubmit={canSubmit}
       />