+++ src/shared/UI/icons/assignable/index.js
@@ -1,0 +1,4 @@
+import AssignableIcon from './component';
+
+export { AssignableIcon };
+export * from './list';