+++ src/shared/UI/inputs/index.js
@@ -1,0 +1,9 @@
+import TextField from './text';
+import CategoryField from './category';
+import ColorField from './color';
+
+export {
+  TextField,
+  CategoryField,
+  ColorField,
+};