+++ src/shared/styles/palette.js
@@ -29,6 +29,7 @@ const colorPalette = colorDefs.map((item) => {
 });
 
 const colorGetterFunctionFor = level => (key) => {
+  if (!key) return null;
   const spectrum = colorPaletteByKey[key] || colorPaletteByKey.gray;
   return spectrum[level];
 };