+++ docs/CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 
+## 1.13.0 - 2020-05-30
+### Added
+* Support for Subject Icons
+
+
+
 ## 1.12.1 - 2020-05-28
 ### Fixed
 * WeekView loading spinner now covers all cases